Almost finished implementation of Assignment 3. Lighting already worked, but now light sources were defined. Z-buffering implemented. Missing interactive change of properties of light and materials.

+++ AlmostGLVisualizer/AlmostGLVisualizer/source/Material.h
@@ -17,7 +17,7 @@ public:
     Vector3f ambient;
     Vector3f diffuse;
     Vector3f specular;
-    float shine;
+    int shine;
 };
